+++ DemoDefragCombiner/DemoDefragCombine.cpp
@@ -800,8 +800,8 @@ qboolean demoCut( const char* outputName, std::vector<std::string>* inputFiles) 
 	demo.cut.Clc.reliableSequence++;
 	demo.cut.Clc.serverMessageSequence++;
 
-	float sourceTime = 0.0f;
-	float time = 10000.0f; // You don't want to start at time 0. It causes incomprehensible weirdness. In fact, it crashes most clients if you try to play back the demo.
+	double sourceTime = 0.0f;
+	double time = 10000.0f; // You don't want to start at time 0. It causes incomprehensible weirdness. In fact, it crashes most clients if you try to play back the demo.
 	float fps = 60.0f;
 	std::map<int, entityState_t> playerEntities;
 	std::map<int, entityState_t> playerEntitiesOld;
@@ -950,8 +950,8 @@ qboolean demoCut( const char* outputName, std::vector<std::string>* inputFiles) 
 			demoCutWriteDeltaSnapshotManual(&commandsToAdd, newHandle, qfalse, &demo.cut.Clc, &demo.cut.Cl, demoType, &playerEntities, &playerEntitiesOld, &mainPlayerPSOld);
 		}
 
-		time += 1000.0f / fps;
-		sourceTime += 1000.0f / fps;
+		time += 1000.0 / fps;
+		sourceTime += 1000.0 / fps;
 		demo.cut.Clc.reliableSequence++;
 		demo.cut.Clc.serverMessageSequence++;
